# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Proteomics Visualization API
  description: Proteomics Visualization Web App API
  version: 1.0.0
paths:
...
  /api/sessions/{session_id}/files/{file_type}/{filename}:
    delete:
      tags:
      - upload
      summary: Delete File
      description: Delete a file from a session.
      operationId: delete_file_api_sessions__session_id__files__file_type___filename__delete
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      - name: file_type
        in: path
        required: true
        schema:
          type: string
          title: File Type
      - name: filename
        in: path
        required: true
        schema:
          type: string
          title: Filename
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/files/tree:
    get:
      tags:
      - files
      summary: List Directory
      description: List directory contents -- folders first, then files.
      operationId: list_directory_api_files_tree_get
      parameters:
      - name: path
        in: query
        required: false
        schema:
          type: string
          description: Relative path to list
          default: ''
          title: Path
        description: Relative path to list
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/files/folders:
    post:
      tags:
      - files
      summary: Create Folder
      description: Create a new folder in the library.
      operationId: create_folder_api_files_folders_post
      requestBody:
        content:
          application/json:
            schema:
              type: object
              title: Body
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/files/upload:
    post:
      tags:
      - files
      summary: Upload Files
      description: Upload files to the library. Only .txt and .csv accepted.
      operationId: upload_files_api_files_upload_post
      parameters:
      - name: target_path
        in: query
        required: false
        schema:
          type: string
          description: Target folder in library
          default: ''
          title: Target Path
        description: Target folder in library
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_files_api_files_upload_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/files/rename:
    put:
      tags:
      - files
      summary: Rename Entry
      description: Rename a file or folder.
      operationId: rename_entry_api_files_rename_put
      requestBody:
        content:
          application/json:
            schema:
              type: object
              title: Body
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/files/move:
    put:
      tags:
      - files
      summary: Move Entry
      description: Move a file or folder to a different parent directory.
      operationId: move_entry_api_files_move_put
      requestBody:
        content:
          application/json:
            schema:
              type: object
              title: Body
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/files/delete:
    delete:
      tags:
      - files
      summary: Delete Entry
      description: Delete a file or folder. Folders are deleted recursively.
      operationId: delete_entry_api_files_delete_delete
      requestBody:
        content:
          application/json:
            schema:
              type: object
              title: Body
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/files/search:
    get:
      tags:
      - files
      summary: Search Files
      description: Search files by name substring.
      operationId: search_files_api_files_search_get
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          description: Search query
          title: Q
        description: Search query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/files/content:
    get:
      tags:
      - files
      summary: Get File Content
      description: Download file contents (for client-side parsing). 10MB max.
      operationId: get_file_content_api_files_content_get
      parameters:
      - name: path
        in: query
        required: true
        schema:
          type: string
          description: Relative path to file
          title: Path
        description: Relative path to file
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/files/select:
    post:
      tags:
      - files
      summary: Select Files For Session
      description: Copy files from library to a session and parse them. Returns ProteomicsFileInfo list.
      operationId: select_files_for_session_api_files_select_post
      requestBody:
        content:
          application/json:
            schema:
              type: object
              title: Body
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_upload_files_api_files_upload_post:
      properties:
        files:
          items:
            type: string
            format: binary
          type: array
          title: Files
      type: object
      required:
      - files
      title: Body_upload_files_api_files_upload_post
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...